requirements.txt
python-dotenv
discord.py
asyncpg
tortoise-orm
Pillow
requests